refactor github indexer a bit
