+++ github-indexer/src/domain/indexers/contributors_projector.rs
