+++ github-indexer/src/domain/indexers/error.rs
