+++ github-indexer/src/domain/indexers/issue/projector.rs
