+++ github-indexer/src/domain/indexers/logged.rs
