+++ github-indexer/src/domain/indexers/pull_request/projector.rs
@@ -6,7 +6,7 @@ use domain::GithubFullPullRequest;
 
 use super::{super::error::Result, Projector};
 use crate::{
-	github_indexer::indexers::contributors_projector::ContributorsProjector,
+	domain::indexers::contributors_projector::ContributorsProjector,
 	models::{ContributionsRepository, GithubPullRequestRepository},
 };
