+++ github-indexer/src/domain/indexers/repo/crawler.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 
 use super::{super::error::Result, IndexedRepo};
 use crate::{
-	github_indexer::indexers::{hash, Crawler},
+	domain::indexers::{hash, Crawler},
 	models::GithubRepoIndexRepository,
 };
