+++ github-indexer/src/domain/indexers/user/projector.rs
