+++ github-indexer/src/domain/mod.rs
@@ -1,0 +1,3 @@
+pub mod controller;
+pub mod indexers;
+pub mod repository;
