+++ github-indexer/src/presentation/cron/mod.rs
@@ -1,6 +1,9 @@
-use anyhow::Result;
+mod scheduler;
 
-use crate::{github_indexer::Scheduler, Config};
+use anyhow::Result;
+pub use scheduler::Scheduler;
+
+use crate::Config;
 
 pub fn bootstrap(config: Config) -> Result<Scheduler> {
 	Scheduler::new(config)
