+++ github-indexer/src/presentation/http/routes/issue.rs
@@ -7,7 +7,7 @@ use presentation::http::guards::ApiKey;
 use rocket::State;
 
 use crate::{
-	github_indexer::indexers::{
+	domain::indexers::{
 		self,
 		optional::{self, Optional},
 		Indexer,
