+++ github-indexer/src/presentation/http/routes/pull_request.rs
@@ -7,7 +7,7 @@ use presentation::http::guards::ApiKey;
 use rocket::State;
 
 use crate::{
-	github_indexer::indexers::{
+	domain::indexers::{
 		self,
 		contributors_projector::ContributorsProjector,
 		optional::{self, Optional},
