+++ github-indexer/src/presentation/mod.rs
@@ -1,12 +1,12 @@
 use anyhow::Result;
 use rocket::{Build, Rocket};
 
-use crate::{github_indexer::Scheduler, Config};
+use crate::Config;
 
 pub mod cron;
 pub mod http;
 
-pub async fn bootstrap(config: Config) -> Result<(Rocket<Build>, Scheduler)> {
+pub async fn bootstrap(config: Config) -> Result<(Rocket<Build>, cron::Scheduler)> {
 	Ok((
 		http::bootstrap(config.clone()).await?,
 		cron::bootstrap(config)?,
